dom/tests/mochitest/mouselock/test_MouseEvents.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
<head>
    <title>Test for Bug 633602 - Mouse Events Allowed and Suppressed.</title>
    <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
    <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
    <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">Mozilla Bug 633602</a>
    <p id="display"></p>
<div id="content">
    <div id="fs">
      <div id="one" style="background-color:Green; width: 100px; height: 100px;">0000</div>
      <div id="outer" style="background-color:Gray; width: 100px; height: 100px;">
        <div id="inner" style="width:50px; height: 100%; background-color:Red; margin:auto;"></div>
    </div>
    <p id="outerResult">00000000</p>
    <p id="innerResult">00000000</p>
</div>
    <pre id="test">
        <script type="application/javascript">
						SpecialPowers.setBoolPref("full-screen-api.allow-trusted-requests-only", false);
						
						// Constant setup
						var RIGHT_MOUSE_CLICK = 2;
						var MIDDLE_MOUSE_CLICK = 1;
						var LEFT_MOUSE_CLICK = 0;
						
						// Variable setup
						var fs = document.getElementById('fs');
						var one = document.getElementById('one');
						var outer = document.getElementById('outer');
						var inner = document.getElementById('inner');
						var outerResult = document.getElementById('outerResult');
						var innerResult = document.getElementById('innerResult');
						setup();
            /** Test for Bug 633602 **/
            SimpleTest.waitForExplicitFinish()
            SimpleTest.waitForFocus(function() {
                setup();
								fs.mozRequestFullScreen();
            });
            				
						document.addEventListener('mozfullscreenchange', mouseCursorRequired, false);

						function mouseCursorRequired() {
								if (document.mozFullScreen) {
										one.innerHTML = "0000";
										outerResult.innerHTML = "00000000";
										innerResult.innerHTML = "00000000";
										if (document.mozFullScreenElement == fs) {
												navigator.pointer.lock(fs);
												synthesizeMouse(fs, 50, 99, { type: 'mousemove' }, window);
												synthesizeMouse(fs, 50, 100, { type: 'mousemove' }, window);
												is(one.innerHTML.substr(0,1), "0", "The event mouseenter should not be firing under Full Screen and Mouse Lock.");
												is(one.innerHTML.substr(1,1), "0", "The event mouseover should not be firing under Full Screen and Mouse Lock.");
												is(one.innerHTML.substr(2,1), "0", "The event mouseleave should not be firing under Full Screen and Mouse Lock.");
												is(one.innerHTML.substr(3,1), "0", "The event mouseout should not be firing under Full Screen and Mouse Lock.");
												document.mozCancelFullScreen();
										}
								} else {
										synthesizeMouse(fs, 50, 99, { type: 'mousemove' }, window);
										synthesizeMouse(fs, 50, 100, { type: 'mousemove' }, window);
										is(one.innerHTML.substr(0,1), "1", "The event mouseenter is not firing under normal circumstances.");
										is(one.innerHTML.substr(1,1), "1", "The event mouseover is not firing under normal circumstances.");
										is(one.innerHTML.substr(2,1), "1", "The event mouseleave is not firing under normal circumstances.");
										is(one.innerHTML.substr(3,1), "1", "The event mouseout is not firing under normal circumstances.");										
										document.removeEventListener('mozfullscreenchange', mouseCursorRequired, false);
										document.addEventListener('mozfullscreenchange', mouseClicks, false);
										outer.mozRequestFullScreen();
								}
						}
										
										
						function mouseClicks() {
							if (document.mozFullScreen) {
									one.innerHTML = "0000";
									outerResult.innerHTML = "00000000";
									innerResult.innerHTML = "00000000";
									if (document.mozFullScreenElement == outer) {
											navigator.pointer.lock(outer);
											sendMouseEvent({ type: 'click' }, inner, window);
											synthesizeMouseAtCenter(outer, { type: 'mousedown', button: LEFT_MOUSE_CLICK }, window);
											synthesizeMouseAtCenter(outer, { type: 'mouseup', button: LEFT_MOUSE_CLICK }, window);
											synthesizeMouseAtCenter(outer, { type: 'mousedown', button: RIGHT_MOUSE_CLICK }, window);
											synthesizeMouseAtCenter(outer, { type: 'mouseup', button: RIGHT_MOUSE_CLICK }, window);
											synthesizeMouseAtCenter(outer, { type: 'mousedown', button: MIDDLE_MOUSE_CLICK }, window);
											synthesizeMouseAtCenter(outer, { type: 'mouseup', button: MIDDLE_MOUSE_CLICK }, window);
											var rect = outer.getBoundingClientRect();
											synthesizeMouseScroll(outer, rect.width/2, rect.height/2, 
													{axis:"vertical", delta:10, type:"DOMMouseScroll"} , window);
											isnot(innerResult.innerHTML.substr(0,1), '1', 
													"The mouseup event for the right mouse button was sent to an element that was not locked.");
											isnot(innerResult.innerHTML.substr(1,1), '1', 
													"The mouseup event for the middle mouse button was sent to an element that was not locked.");
											isnot(innerResult.innerHTML.substr(2,1), '1', 
													"The mouseup event for the left mouse button was sent to an element that was not locked.");
											isnot(innerResult.innerHTML.substr(3,1), '1', 
													"The mousedown event for the right mouse button was sent to an element that was not locked.");
											isnot(innerResult.innerHTML.substr(4,1), '1', 
													"The mousedown event for the middle mouse button was sent to an element that was not locked.");
											isnot(innerResult.innerHTML.substr(5,1), '1', 
													"The mousedown event for the left mouse button was sent to an element that was not locked.");
											isnot(innerResult.innerHTML.substr(6,1), '1', 
													"The DOMMouseScroll event was sent to an element that was not locked.");
											isnot(innerResult.innerHTML.substr(7,1), '1', 
													"The click event  was sent to an element that was not locked.");

											is(outerResult.innerHTML.substr(0,1), '1', 
													"The full screened mouse locked element did not get the mouseup event for the right mouse button.");
											is(outerResult.innerHTML.substr(1,1), '1', 
													"The full screened mouse locked element did not get the mouseup event for the middle mouse button.");
											is(outerResult.innerHTML.substr(2,1), '1', 
													"The full screened mouse locked element did not get the mouseup event for the left mouse button.");
											is(outerResult.innerHTML.substr(3,1), '1', 
													"The full screened mouse locked element did not get the mousedown event for the right mouse button.");
											is(outerResult.innerHTML.substr(4,1), '1', 
													"The full screened mouse locked element did not get the mousedown event for the middle mouse button.");
											is(outerResult.innerHTML.substr(5,1), '1', 
													"The full screened mouse locked element did not get the mousedown event for the left mouse button.");
											is(outerResult.innerHTML.substr(6,1), '1', 
													"The full screened mouse locked element did not get the DOMMouseScroll event.");
											is(outerResult.innerHTML.substr(7,1), '1', 
													"The full screened mouse locked element did not get the click event.");     
											document.mozCancelFullScreen();
									}
							} else {
								var rect = outer.getBoundingClientRect();								
                sendMouseEvent({ type: 'click' }, inner, window);
								synthesizeMouseAtCenter(outer, { type: 'mousedown', button: LEFT_MOUSE_CLICK }, window);
								synthesizeMouseAtCenter(outer, { type: 'mouseup', button: LEFT_MOUSE_CLICK }, window);
                synthesizeMouseAtCenter(outer, { type: 'mousedown', button: RIGHT_MOUSE_CLICK }, window);
								synthesizeMouseAtCenter(outer, { type: 'mouseup', button: RIGHT_MOUSE_CLICK }, window);
                synthesizeMouseAtCenter(outer, { type: 'mousedown', button: MIDDLE_MOUSE_CLICK }, window);
								synthesizeMouseAtCenter(outer, { type: 'mouseup', button: MIDDLE_MOUSE_CLICK }, window);
                synthesizeMouseScroll(outer, rect.width/2, rect.height/2, 
                    {axis:"vertical", delta:10, type:"DOMMouseScroll"} , window);
                is(innerResult.innerHTML.substr(0,1), '1', 
                    "The mouseup event for the right mouse button did not fire for inner element under normal circumstances.");
                is(innerResult.innerHTML.substr(1,1), '1', 
                    "The mouseup event for the middle mouse button did not fire for inner element under normal circumstances.");
                is(innerResult.innerHTML.substr(2,1), '1', 
                    "The mouseup event for the left mouse button did not fire for inner element under normal circumstances.");
                is(innerResult.innerHTML.substr(3,1), '1', 
                    "The mousedown event for the right mouse button did not fire for inner element under normal circumstances.");
                is(innerResult.innerHTML.substr(4,1), '1', 
                    "The mousedown event for the middle mouse button did not fire for inner element under normal circumstances.");
                is(innerResult.innerHTML.substr(5,1), '1', 
                    "The mousedown event for the left mouse button did not fire for inner element under normal circumstances.");
                is(innerResult.innerHTML.substr(6,1), '1', 
                    "The DOMMouseScroll event did not fire for inner element under normal circumstances.");
                is(innerResult.innerHTML.substr(7,1), '1', 
                    "The click event did not fire for inner element under normal circumstances.");

                is(outerResult.innerHTML.substr(0,1), '1', 
                    "The mouseup event for the right mouse button did not bubble up to the outer element under normal circumstances.");
                is(outerResult.innerHTML.substr(1,1), '1', 
                    "The mouseup event for the middle mouse button did not bubble up to the outer element under normal circumstances.");
                is(outerResult.innerHTML.substr(2,1), '1', 
                    "The mouseup event for the left mouse button did not bubble up to the outer element under normal circumstances.");
                is(outerResult.innerHTML.substr(3,1), '1', 
                    "The mousedown event for the right mouse button did not bubble up to the outer element under normal circumstances.");
                is(outerResult.innerHTML.substr(4,1), '1', 
                    "The mousedown event for the middle mouse button did not bubble up to the outer element under normal circumstances.");
                is(outerResult.innerHTML.substr(5,1), '1', 
                    "The mousedown event for the left mouse button did not bubble up to the outer element under normal circumstances.");
                is(outerResult.innerHTML.substr(6,1), '1', 
                    "The DOMMouseScroll event did not bubble up to the outer element under normal circumstances.");
                is(outerResult.innerHTML.substr(7,1), '1', 
                    "The click event did not bubble up to the outer element under normal circumstances.");
								SimpleTest.finish();
							}
						}
            
            function setup() {    
                
                // Events that only exist because a mouse exists
                one.addEventListener('mouseout', function() { one.innerHTML = one.innerHTML.substr(0,3) + '1'; } , false);
                one.addEventListener('mouseleave', function() { one.innerHTML = one.innerHTML.substr(0,2) + '1' + one.innerHTML.substr(3); } , false);
                one.addEventListener('mouseover', function() { one.innerHTML = one.innerHTML.substr(0,1) + '1' + one.innerHTML.substr(2); } , false);
                one.addEventListener('mouseenter', function() { one.innerHTML = '1' + one.innerHTML.substr(1); } , false);
                
                // Events that should be there regardless of mouse state
                outer.addEventListener('click', function() { outerResult.innerHTML = outerResult.innerHTML.substr(0,7) + '1'; },false);
                outer.addEventListener('DOMMouseScroll', function() { outerResult.innerHTML = outerResult.innerHTML.substr(0,6) + '1' + outerResult.innerHTML.substr(7); },false);
                
                outer.addEventListener('mousedown', function(e) { 
                  if (e) {
                    switch (e.button) {
                      case LEFT_MOUSE_CLICK:
                        outerResult.innerHTML = outerResult.innerHTML.substr(0,5) + '1' + outerResult.innerHTML.substr(6);
                        break;
                      case MIDDLE_MOUSE_CLICK:
                         outerResult.innerHTML = outerResult.innerHTML.substr(0,4) + '1' + outerResult.innerHTML.substr(5);
                         break;
                      case RIGHT_MOUSE_CLICK:
                        outerResult.innerHTML = outerResult.innerHTML.substr(0,3) + '1' + outerResult.innerHTML.substr(4);
                        break;
                    }
                  }
                },false);
                
                outer.addEventListener('mouseup', function(e) { 
                  if (e) {
                      switch (e.button) {
                        case LEFT_MOUSE_CLICK:
                          outerResult.innerHTML = outerResult.innerHTML.substr(0,2) + '1' + outerResult.innerHTML.substr(3);
                          break;
                        case MIDDLE_MOUSE_CLICK:
                           outerResult.innerHTML = outerResult.innerHTML.substr(0,1) + '1' + outerResult.innerHTML.substr(2);
                           break;
                        case RIGHT_MOUSE_CLICK:
                          outerResult.innerHTML = '1' + outerResult.innerHTML.substr(1);
                          break;
                      }
                    }
                },false);

                inner.addEventListener('click', function() { innerResult.innerHTML = innerResult.innerHTML.substr(0,7) + '1'; },false);
                inner.addEventListener('DOMMouseScroll', function() { innerResult.innerHTML = innerResult.innerHTML.substr(0,6) + '1' + innerResult.innerHTML.substr(7); },false);

                inner.addEventListener('mousedown', function(e) { 
                  if (e) {
                    switch (e.button) {
                      case LEFT_MOUSE_CLICK:
                        innerResult.innerHTML = innerResult.innerHTML.substr(0,5) + '1' + innerResult.innerHTML.substr(6);
                        break;
                      case MIDDLE_MOUSE_CLICK:
                         innerResult.innerHTML = innerResult.innerHTML.substr(0,4) + '1' + innerResult.innerHTML.substr(5);
                         break;
                      case RIGHT_MOUSE_CLICK:
                        innerResult.innerHTML = innerResult.innerHTML.substr(0,3) + '1' + innerResult.innerHTML.substr(4);
                        break;
                    }
                  }
                },false);
                inner.addEventListener('mouseup', function(e) { 
                  if (e) {
                      switch (e.button) {
                        case LEFT_MOUSE_CLICK:
                          innerResult.innerHTML = innerResult.innerHTML.substr(0,2) + '1' + innerResult.innerHTML.substr(3);
                          break;
                        case MIDDLE_MOUSE_CLICK:
                           innerResult.innerHTML = innerResult.innerHTML.substr(0,1) + '1' + innerResult.innerHTML.substr(2);
                           break;
                        case RIGHT_MOUSE_CLICK:
                          innerResult.innerHTML = '1' + innerResult.innerHTML.substr(1);
                          break;
                      }
                  }
                },false);    
            }
        </script>
    </pre>
</body>
</html>
